+++ crates/shrs/Cargo.toml
@@ -35,6 +35,7 @@ shrs_core = { path = "../shrs_core", version = "^0.0.2" }
 shrs_line = { path = "../shrs_line", version = "^0.0.2" }
 shrs_lang = { path = "../shrs_lang", version = "^0.0.2" }
 shrs_job = { path = "../shrs_job", version = "^0.0.2" }
+shrs_utils = { path = "../shrs_utils", version = "^0.0.2" }
 
 serde = { version = "1", features = ["derive"], optional = true }
