+++ crates/shrs_utils/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod algo;
 pub mod cursor_buffer;
 pub mod macros;
+pub mod styled_buf;
